requirements.txt
Django==2.1.7
django-rest-framework==0.1.0
djangorestframework==3.9.1
djangorestframework-gis==0.14
mysqlclient==1.4.2.post1
Pillow==5.4.1
pytz==2018.9
